+++ unstructured_inference/inference/layout.py
@@ -12,7 +12,8 @@ from pdfminer import psparser
 from pdfminer.high_level import extract_pages
 from PIL import Image, ImageSequence
 from pytesseract import Output
-from unstructured_inference.models import paddle_ocr
+
+from unstructured_inference.constants import OCRMode
 from unstructured_inference.inference.elements import (
     EmbeddedTextRegion,
     ImageTextRegion,
@@ -83,9 +84,10 @@ class DocumentLayout:
         fixed_layouts: Optional[List[Optional[List[TextRegion]]]] = None,
         ocr_strategy: str = "auto",
         ocr_languages: str = "eng",
-        ocr_mode: str = "entire_page",
+        ocr_mode: str = OCRMode.FULL_PAGE.value,
         extract_tables: bool = False,
         pdf_image_dpi: int = 200,
+        **kwargs,
     ) -> DocumentLayout:
         """Creates a DocumentLayout from a pdf file."""
         logger.info(f"Reading PDF for file: {filename} ...")
@@ -125,6 +127,7 @@ class DocumentLayout:
                         ocr_mode=ocr_mode,
                         fixed_layout=fixed_layout,
                         extract_tables=extract_tables,
+                        **kwargs,
                     )
                     pages.append(page)
             return cls.from_pages(pages)
@@ -137,9 +140,10 @@ class DocumentLayout:
         element_extraction_model: Optional[UnstructuredElementExtractionModel] = None,
         ocr_strategy: str = "auto",
         ocr_languages: str = "eng",
-        ocr_mode: str = "entire_page",
+        ocr_mode: str = OCRMode.FULL_PAGE.value,
         fixed_layout: Optional[List[TextRegion]] = None,
         extract_tables: bool = False,
+        **kwargs,
     ) -> DocumentLayout:
         """Creates a DocumentLayout from an image file."""
         logger.info(f"Reading image file: {filename} ...")
@@ -167,8 +171,10 @@ class DocumentLayout:
                 layout=None,
                 ocr_strategy=ocr_strategy,
                 ocr_languages=ocr_languages,
+                ocr_mode=ocr_mode,
                 fixed_layout=fixed_layout,
                 extract_tables=extract_tables,
+                **kwargs,
             )
             pages.append(page)
         return cls.from_pages(pages)
@@ -189,8 +195,10 @@ class PageLayout:
         element_extraction_model: Optional[UnstructuredElementExtractionModel] = None,
         ocr_strategy: str = "auto",
         ocr_languages: str = "eng",
-        ocr_mode: str = "entire_page",
+        ocr_mode: str = OCRMode.FULL_PAGE.value,
         extract_tables: bool = False,
+        analysis: bool = False,
+        supplement_with_ocr_elements: bool = True,
     ):
         if detection_model is not None and element_extraction_model is not None:
             raise ValueError("Only one of detection_model and extraction_model should be passed.")
@@ -212,6 +220,10 @@ class PageLayout:
         self.ocr_languages = ocr_languages
         self.ocr_mode = ocr_mode
         self.extract_tables = extract_tables
+        self.analysis = analysis
+        self.inferred_layout: Optional[List[LayoutElement]] = None
+        self.ocr_layout: Optional[List[TextRegion]] = None
+        self.supplement_with_ocr_elements = supplement_with_ocr_elements
 
     def __str__(self) -> str:
         return "\n\n".join([str(element) for element in self.elements])
@@ -248,9 +260,9 @@ class PageLayout:
         # remote call in the future.
         inferred_layout: List[LayoutElement] = self.detection_model(self.image)
 
-        if self.ocr_mode == "individual_blocks":
+        if self.ocr_mode == OCRMode.INDIVIDUAL_BLOCKS.value:
             ocr_layout = None
-        elif self.ocr_mode == "entire_page":
+        elif self.ocr_mode == OCRMode.FULL_PAGE.value:
             ocr_layout = None
             try:
                 paddle_result = paddle_ocr.load_agent().ocr(np.array(self.image), cls=True)
@@ -269,10 +281,12 @@ class PageLayout:
                 and "R_50" not in self.detection_model.model_path
             ):
                 threshold_kwargs = {"same_region_threshold": 0.5, "subregion_threshold": 0.5}
-            inferred_layout = merge_inferred_layout_with_extracted_layout(
+            merged_layout = merge_inferred_layout_with_extracted_layout(
                 inferred_layout=inferred_layout,
                 extracted_layout=self.layout,
+                page_image_size=self.image.size,
                 ocr_layout=ocr_layout,
+                supplement_with_ocr_elements=self.supplement_with_ocr_elements,
                 **threshold_kwargs,
             )
         elif ocr_layout is not None:
@@ -284,13 +298,20 @@ class PageLayout:
                 and "R_50" not in self.detection_model.model_path
             ):
                 threshold_kwargs = {"subregion_threshold": 0.3}
-            inferred_layout = merge_inferred_layout_with_ocr_layout(
+            merged_layout = merge_inferred_layout_with_ocr_layout(
                 inferred_layout=inferred_layout,
                 ocr_layout=ocr_layout,
+                supplement_with_ocr_elements=self.supplement_with_ocr_elements,
                 **threshold_kwargs,
             )
+        else:
+            merged_layout = inferred_layout
 
-        elements = self.get_elements_from_layout(cast(List[TextRegion], inferred_layout))
+        elements = self.get_elements_from_layout(cast(List[TextRegion], merged_layout))
+
+        if self.analysis:
+            self.inferred_layout = inferred_layout
+            self.ocr_layout = ocr_layout
 
         if inplace:
             self.elements = elements
@@ -328,6 +349,7 @@ class PageLayout:
         self,
         colors: Optional[Union[List[str], str]] = None,
         image_dpi: int = 200,
+        annotation_data: Optional[dict[str, dict]] = None,
     ) -> Image.Image:
         """Annotates the elements on the page image."""
         if colors is None:
@@ -347,9 +369,18 @@ class PageLayout:
         else:
             img = self._get_image(self.document_filename, self.number, image_dpi)
 
-        for el, color in zip(self.elements, colors):
-            if isinstance(el, Rectangle):
-                img = draw_bbox(img, el, color=color)
+        if annotation_data is None:
+            for el, color in zip(self.elements, colors):
+                if isinstance(el, Rectangle):
+                    img = draw_bbox(img, el, color=color)
+        else:
+            for attribute, style in annotation_data.items():
+                if hasattr(self, attribute) and getattr(self, attribute):
+                    color = style["color"]
+                    width = style["width"]
+                    for region in getattr(self, attribute):
+                        if isinstance(region, Rectangle):
+                            img = draw_bbox(img, region, color=color, width=width)
 
         return img
 
@@ -384,11 +415,15 @@ class PageLayout:
         layout: Optional[List[TextRegion]] = None,
         ocr_strategy: str = "auto",
         ocr_languages: str = "eng",
-        ocr_mode: str = "entire_page",
+        ocr_mode: str = OCRMode.FULL_PAGE.value,
         extract_tables: bool = False,
         fixed_layout: Optional[List[TextRegion]] = None,
+        **kwargs,
     ):
         """Creates a PageLayout from an already-loaded PIL Image."""
+        analysis = kwargs.get("analysis", False)
+        supplement_with_ocr_elements = kwargs.get("supplement_with_ocr_elements", True)
+
         page = cls(
             number=number,
             image=image,
@@ -399,6 +434,8 @@ class PageLayout:
             ocr_languages=ocr_languages,
             ocr_mode=ocr_mode,
             extract_tables=extract_tables,
+            analysis=analysis,
+            supplement_with_ocr_elements=supplement_with_ocr_elements,
         )
         if page.element_extraction_model is not None:
             page.get_elements_using_image_extraction()
@@ -428,10 +465,11 @@ def process_data_with_model(
     is_image: bool = False,
     ocr_strategy: str = "auto",
     ocr_languages: str = "eng",
-    ocr_mode: str = "entire_page",
+    ocr_mode: str = OCRMode.FULL_PAGE.value,
     fixed_layouts: Optional[List[Optional[List[TextRegion]]]] = None,
     extract_tables: bool = False,
     pdf_image_dpi: Optional[int] = None,
+    **kwargs,
 ) -> DocumentLayout:
     """Processes pdf file in the form of a file handler (supporting a read method) into a
     DocumentLayout by using a model identified by model_name."""
@@ -448,6 +486,7 @@ def process_data_with_model(
             fixed_layouts=fixed_layouts,
             extract_tables=extract_tables,
             pdf_image_dpi=pdf_image_dpi,
+            **kwargs,
         )
 
     return layout
@@ -459,10 +498,11 @@ def process_file_with_model(
     is_image: bool = False,
     ocr_strategy: str = "auto",
     ocr_languages: str = "eng",
-    ocr_mode: str = "entire_page",
+    ocr_mode: str = OCRMode.FULL_PAGE.value,
     fixed_layouts: Optional[List[Optional[List[TextRegion]]]] = None,
     extract_tables: bool = False,
     pdf_image_dpi: Optional[int] = None,
+    **kwargs,
 ) -> DocumentLayout:
     """Processes pdf file with name filename into a DocumentLayout by using a model identified by
     model_name."""
@@ -493,6 +533,7 @@ def process_file_with_model(
             ocr_languages=ocr_languages,
             ocr_mode=ocr_mode,
             extract_tables=extract_tables,
+            **kwargs,
         )
         if is_image
         else DocumentLayout.from_file(
@@ -505,6 +546,7 @@ def process_file_with_model(
             fixed_layouts=fixed_layouts,
             extract_tables=extract_tables,
             pdf_image_dpi=pdf_image_dpi,
+            **kwargs,
         )
     )
     return layout